refactor(alerts): rename redesigned alert modal
Generated-By: PostHog Code
Task-Id: 38d68d83-d9f6-49d8-af12-3d8ab8f46aab
